+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.0.14
 
-1. Fix `Response` and `Request` clone failure.
+1. fix `Response` and `Request` clone failure.
+2. fix that the response fully complies with [MDN Response](https://developer.mozilla.org/zh-CN/docs/Web/API/Response) conventions, and now all attributes are read-only
 
 # 0.0.13